+++ KnockKnock-iOS/Scenes/My/Views/MyView.swift
@@ -56,9 +56,9 @@ final class MyView: UIView {
     $0.textColor = KKDS.Color.green50
   }
 
-  let alertView = AlertView().then {
-    $0.isHidden = true
-  }
+//  let alertView = AlertView().then {
+//    $0.isHidden = true
+//  }
 
   // MARK: - Initialize
   
@@ -81,34 +81,13 @@ final class MyView: UIView {
   func setNickname(nickname: String) {
     self.userNameLabel.text = "반가워요 \(nickname)님 🌿"
   }
-
-  /// menuType == nil 일 때 alertView hidden
-  func setAlertView(menuType: MyMenuType?) {
-    var content = ""
-
-    guard let menuType = menuType else {
-      self.alertView.isHidden = true
-      return
-    }
-    
-    if menuType == .signOut {
-      content = "계정을 삭제하면 게시글, 좋아요, 댓글 등 모든 활동 정보가 삭제됩니다. 그래도 탈퇴 하시겠습니까?"
-      self.alertView.bind(content: content, isCancelActive: true)
-
-    } else if menuType == .versionInfo {
-      content = "현재 최신버전을 사용중입니다."
-      self.alertView.bind(content: content, isCancelActive: false)
-
-    }
-    self.alertView.isHidden = false
-  }
   
   // MARK: - Constraints
   
   private func setupConstraints() {
     [self.myTableView].addSubViews(self)
     [self.loginButton, self.userNameLabel].addSubViews(self.myTableHeaderView)
-    [self.alertView].addSubViews(self)
+//    [self.alertView].addSubViews(self)
     
     self.myTableView.snp.makeConstraints {
       $0.edges.equalTo(self.safeAreaLayoutGuide)
@@ -123,9 +102,9 @@ final class MyView: UIView {
       $0.leading.trailing.equalToSuperview().inset(Metric.userNameLabelLeadingMargin)
       $0.centerY.equalToSuperview()
     }
-
-    self.alertView.snp.makeConstraints {
-      $0.edges.equalToSuperview()
-    }
+//
+//    self.alertView.snp.makeConstraints {
+//      $0.edges.equalToSuperview()
+//    }
   }
 }
